pyproject.toml
[tool.poetry]
name = "personal-assistant"
version = "0.1.0"
description = "A personal assistant with RAG capabilities using PostgreSQL/pgvector"
authors = ["Wolf Bubberman <wolfbubberman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
click = "^8.1.0"
requests = "^2.31.0"
anthropic = "^0.3.0"
langchain = "^0.1.0"
openai = "^1.12.0"
tiktoken = "^0.9.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.0.0"
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
python-multipart = "^0.0.6"
unstructured = "^0.10.30"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.0"                                      # Using specific version range due to binary nature
pgvector = "^0.2.0"
alembic = "^1.13.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
aiosmtplib = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
